requirements.txt
asgiref==3.7.2
boto3==1.26.59
botocore==1.29.165
Django==5.0.2
django-storages==1.14.2
djangorestframework==3.15.1
jmespath==1.0.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==1.26.18
